// src/pages/ui/generated/catalog-api/resource-list.ts
// Generated by scripts/catalog-api.ts. Do not edit.

export const API = [
	{
		name: "ResourceList",
		props: [
			{
				name: "title",
				type: "string",
				required: true,
			},
			{
				name: "description",
				type: "string",
				required: false,
			},
			{
				name: "data",
				type: "ResourceListRow[]",
				required: true,
			},
		],
	},
];